Remove user from party when account is deleted
